Added confirmation buttons to purge

+++ commands/admin.py
@@ -24,6 +24,13 @@ from helpers.constants import (
     GUILD_PROD,
     ROLE_TIME_OUT,
     CHAN_IOWA,
+    CAT_GAMEDAY,
+    CAT_GENERAL,
+    ROLE_EVERYONE_PROD,
+    CHAN_HYPE_GROUP,
+    CHAN_GENERAL,
+    CHAN_DISCUSSION_LIVE,
+    CHAN_DISCUSSION_STREAMING,
 )
 from helpers.embed import buildEmbed
 from helpers.misc import discordURLFormatter
@@ -33,8 +40,135 @@ from objects.Exceptions import CommandException, UserInputException
 logger = logging.getLogger(__name__)
 
 
-async def validate_purge(interaction: discord.Interaction):
-    ...
+async def alert_gameday_channels(client: discord.ext.commands.Bot, on: bool):
+    chan_general = await client.fetch_channel(CHAN_GENERAL)
+    chan_live = await client.fetch_channel(CHAN_DISCUSSION_LIVE)
+    chan_streaming = await client.fetch_channel(CHAN_DISCUSSION_STREAMING)
+
+    if on:
+        embed = buildEmbed(
+            title="Game Day Mode",
+            description="Game day mode is now on!",
+            fields=[
+                dict(
+                    name="Live TV",
+                    value=f"{chan_live.mention} text and voice channels are for users who are watching live.",
+                    inline=False,
+                ),
+                dict(
+                    name="Streaming",
+                    value=f"{chan_streaming.mention} text and voice channels are for users who are streaming the game.",
+                    inline=False,
+                ),
+                dict(
+                    name="Info",
+                    value="All channels in the Huskers category will be turned off until the game day mode is disabled.",
+                    inline=False,
+                ),
+            ],
+        )
+    else:
+        embed = buildEmbed(
+            title="Game Day Mode",
+            description="Game day mode is now off!",
+            fields=[
+                dict(
+                    name="Info",
+                    value=f"Game day channels have been disabled and General categories channels have been enabled. Regular discussion may continue in {chan_general.mention}.",
+                    inline=False,
+                )
+            ],
+        )
+
+    await chan_general.send(embed=embed)
+    await chan_live.send(embed=embed)
+    await chan_streaming.send(embed=embed)
+
+
+async def process_gameday(mode: bool, guild: discord.Guild):
+    gameday_category = guild.get_channel(CAT_GAMEDAY)
+    general_category = guild.get_channel(CAT_GENERAL)
+    everyone = guild.get_role(ROLE_EVERYONE_PROD)
+
+    logger.info(f"Creating permissions to be [{mode}]")
+
+    perms_text = discord.PermissionOverwrite()
+
+    perms_text.view_channel = mode  # noqa
+    perms_text.send_messages = mode  # noqa
+    perms_text.read_messages = mode  # noqa
+
+    perms_text_opposite = discord.PermissionOverwrite()
+    perms_text_opposite.send_messages = not mode  # noqa
+
+    perms_voice = discord.PermissionOverwrite()
+    perms_voice.view_channel = mode  # noqa
+    perms_voice.connect = mode  # noqa
+    perms_voice.speak = mode  # noqa
+
+    logger.info(f"Permissions created")
+
+    for channel in general_category.channels:
+
+        if channel.id in CHAN_HYPE_GROUP:
+            continue
+
+        try:
+            logger.info(
+                f"Attempting to changes permissions for [{channel}] to [{not mode}]"
+            )
+            if channel.type == discord.ChannelType.text:
+                await channel.set_permissions(everyone, overwrite=perms_text_opposite)
+        except:  # noqa
+            logger.info(f"Unable to change permissions for [{channel}] to [{not mode}]")
+            continue
+
+        logger.info(f"Changed permissions for [{channel}] to [{not mode}]")
+
+    for channel in gameday_category.channels:
+        try:
+            logger.info(
+                f"Attempting to changes permissions for [{channel}] to [{mode}]"
+            )
+
+            if channel.type == discord.ChannelType.text:
+                await channel.set_permissions(everyone, overwrite=perms_text)
+            elif channel.type == discord.ChannelType.voice:
+                await channel.set_permissions(everyone, overwrite=perms_voice)
+
+                # TODO Trying to kick people from voice channels if game day mode is turned off.
+            else:
+                logger.info(f"Unable to change permissions for [{channel}] to [{mode}]")
+                continue
+            logger.info(f"Changed permissions for [{channel}] to [{mode}]")
+        except discord.errors.Forbidden:
+            raise CommandException(
+                "The bot does not have access to change permissions!"
+            )
+        except:  # noqa
+            continue
+
+    logger.info(f"All permissions changes applied")
+
+
+class Confirm(discord.ui.View):
+    def __init__(self):
+        super(Confirm, self).__init__()
+        self.value = None
+
+    @discord.ui.button(label="Confirm", style=discord.ButtonStyle.green)
+    async def confirm(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ) -> None:
+        await interaction.response.send_message("Confirming", ephemeral=True)
+        self.value = True
+        self.stop()
+
+    @discord.ui.button(label="Cancel", style=discord.ButtonStyle.grey)
+    async def cancel(self, interaction: discord.Interaction, button: discord.ui.Button):
+        await interaction.response.send_message("Cancelling", ephemeral=True)
+        self.value = False
+        self.stop()
 
 
 async def college_purge_messages(channel: Any, all_messages: bool = False):
@@ -62,6 +196,24 @@ async def college_purge_messages(channel: Any, all_messages: bool = False):
         )
 
     return msgs
+
+
+async def confirm_purge(interaction: discord.Interaction):
+
+    view = Confirm()
+    await interaction.response.send_message(
+        "Do you want to continue?", view=view, ephemeral=True
+    )
+    await view.wait()
+
+    if view.value is None:
+        logger.info("Purge confirmation timed out!")
+    elif view.value:
+        logger.info("Purge confirmed. Continuing!")
+        return True
+    else:
+        logger.info("Purge cancelled. Halting!")
+        return False
 
 
 class AdminCog(commands.Cog, name="Admin Commands"):
@@ -203,39 +355,41 @@ class AdminCog(commands.Cog, name="Admin Commands"):
         name="bot", description="Purge the 100 most recent bot messages"
     )
     async def purge_bot(self, interaction: discord.Interaction) -> None:
-        # TODO Add a "double check" button to make sure you want to delete
-        await validate_purge(interaction)
-
         assert type(interaction.channel) is discord.TextChannel, CommandException(
             "Unable to run this command outside text channels."
         )
-        await interaction.response.defer(ephemeral=True)
+
+        if not await confirm_purge(interaction):
+            return
+
         msgs = await college_purge_messages(
             channel=interaction.channel, all_messages=False
         )
+
         await interaction.channel.delete_messages(msgs)
-        logger.info(f"Bulk delete of {len(msgs)} messages successful.")
         await interaction.followup.send(
             f"Bulk delete of {len(msgs)} messages successful.", ephemeral=True
         )
+        logger.info(f"Bulk delete of {len(msgs)} messages successful.")
 
     @group_purge.command(name="all", description="Purge the 100 most recent messages")
     async def purge_all(self, interaction: discord.Interaction) -> None:
-        # TODO Add a "double check" button to make sure you want to delete
-        await validate_purge(interaction)
-
         assert type(interaction.channel) is discord.TextChannel, CommandException(
             "Unable to run this command outside text channels."
         )
-        await interaction.response.defer(ephemeral=True)
+
+        if not await confirm_purge(interaction):
+            return
+
         msgs = await college_purge_messages(
             channel=interaction.channel, all_messages=True
         )
+
         await interaction.channel.delete_messages(msgs)
-        logger.info(f"Bulk delete of {len(msgs)} messages successful.")
         await interaction.followup.send(
             f"Bulk delete of {len(msgs)} messages successful.", ephemeral=True
         )
+        logger.info(f"Bulk delete of {len(msgs)} messages successful.")
 
     @app_commands.command(name="quit", description="Turn off Bot Frost")
     @app_commands.default_permissions(manage_messages=True)
@@ -472,21 +626,32 @@ class AdminCog(commands.Cog, name="Admin Commands"):
     @group_gameday.command(
         name="on",
         description="WIP: Turn game day mode on. Restricts access to server channels.",
-    )  # TODO
+    )
+    @app_commands.default_permissions(manage_messages=True)
     async def gameday_on(self, interaction: discord.Interaction) -> None:
-        ...
+        logger.info(f"Game Day: On")
+        await interaction.response.defer(ephemeral=True)
+        await interaction.followup.send("Processing!")
+        await process_gameday(True, interaction.guild)
+        await alert_gameday_channels(self.client, True)
 
     @group_gameday.command(
         name="off",
         description="WIP: Turn game day mode off. Restores access to server channels.",
-    )  # TODO
+    )
+    @app_commands.default_permissions(manage_messages=True)
     async def gameday_off(self, interaction: discord.Interaction) -> None:
-        ...
+        logger.info(f"Game Day: Off")
+        await interaction.response.defer(ephemeral=True)
+        await interaction.followup.send("Processing!")
+        await process_gameday(False, interaction.guild)
+        await alert_gameday_channels(self.client, False)
 
     @app_commands.command(name="smms")  # TODO
     @app_commands.default_permissions(manage_messages=True)
     async def smms(self, interaction: discord.Interaction) -> None:
-        ...
+        await interaction.response.defer(ephemeral=True)
+        await interaction.followup.send("Hehe")
 
 
 async def setup(bot: commands.Bot) -> None:
